refactor(shared): extract _normalize_user into shared.users
Open WebUI may pass raw user dictionaries into tool contexts. Share the normalizer used before middleware calls instead of keeping per-plugin copies.
