+++ tools/sub_agent.py
@@ -304,6 +304,23 @@ TERMINAL_EVENT_TOOLS: set[str] = {
     "run_command",
 }
 
+# --- inlined from src/owui_ext/shared/users.py (owui_ext.shared.users) ---
+from typing import Any
+def _normalize_user(user: Any) -> Any:
+    """Convert raw __user__ dict payloads into UserModel when needed."""
+    if user is None or hasattr(user, "id"):
+        return user
+    if isinstance(user, dict):
+        try:
+            from open_webui.models.users import UserModel
+
+            return UserModel(**user)
+        except Exception:
+            from types import SimpleNamespace
+
+            return SimpleNamespace(**user)
+    return user
+
 log = logging.getLogger(__name__)
 
 _core_process_tool_result = None
@@ -637,22 +654,6 @@ def normalize_terminal_tools_result(*, terminal_tools_result: Any, extra_params:
     if isinstance(terminal_tools, dict):
         return terminal_tools
     return {}
-
-
-def _normalize_user(user: Any) -> Any:
-    """Convert raw __user__ dict payloads into UserModel when needed."""
-    if user is None or hasattr(user, "id"):
-        return user
-    if isinstance(user, dict):
-        try:
-            from open_webui.models.users import UserModel
-
-            return UserModel(**user)
-        except Exception:
-            from types import SimpleNamespace
-
-            return SimpleNamespace(**user)
-    return user
 
 
 async def resolve_mcp_tools(
